Add format method to lib/url.
This method allows for formatting a URL into other types, like
turning an absolute URL into a scheme-relative one. While formatting, it
will also normalize the URL.

It will not format a lesser type into a more complete type, as it's not
possible to do so without adding extra information.

It does not handle path-relative URLs.